requirements.txt
flask>=3.0
pymongo>=4.6
pydantic>=2.0
python-dotenv>=1.0
spacy>=3.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
